+++ src/main/kotlin/cc/polyfrost/polyui/layout/impl/extension/DraggableLayout.kt
@@ -43,7 +43,7 @@ class DraggableLayout(layout: Layout) : PointerLayout(layout) {
 
     override fun accept(event: Events): Boolean {
         if (event is Events.MousePressed) {
-            if (event.mods != 0.toShort()) return super.accept(event)
+            if (event.mods.toInt() != 0) return super.accept(event)
             mouseClickX = polyui.eventManager.mouseX - ptr.at.a.px
             mouseClickY = polyui.eventManager.mouseY - ptr.at.b.px
             mouseDown = components.noneAre { it.mouseOver } && children.noneAre { it.mouseOver }
